refactor(core): extract LibraryBook DTOs + injectable cover/file URL layer
Move the frontend-facing DTOs (LibraryBook, LibraryAuthor, LocalOrRemoteUrl,
the file list, Identifier) and the libcalibre->DTO conversion out of the Tauri
crate into a new platform-agnostic `citadel-core` crate, so a future
citadel-server can serialize the same shapes over HTTP.

The cover/file URL scheme is now injected via a `BookUrlBuilder` trait instead
of being hardcoded to Tauri's asset:// protocol:

- AssetUrlBuilder reproduces the desktop app's asset://localhost/... output
  byte-for-byte, including the ?v=<mtime_ms> cache-bust and has_cover trust.
- HttpUrlBuilder adds the Remote { url } path (https://.../api/books/<id>/...)
  for a future HTTP caller.

citadel-core depends only on libcalibre/serde/specta/urlencoding -- no tauri.
The Tauri shell re-exports LibraryBook/LibraryAuthor from citadel-core, injects
AssetUrlBuilder, and cover_thumbs now shares citadel_core::path_to_asset_url
(libs/util.rs removed). A unit test exercises both builders over the same Book.

bindings.ts is unchanged (specta emits identical names/shapes).
