+++ Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
 resolver = "2"
-members = ["crates/libcalibre", "src-tauri"]
+members = ["crates/citadel-core", "crates/libcalibre", "src-tauri"]
 
 # Cover thumbnail generation decodes/resizes JPEGs through these crates;
 # unoptimized they run 20-50x slower, turning the background thumbnail warm
