+++ crates/citadel-core/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "citadel-core"
+version = "0.1.0"
+edition = "2021"
+
+# Platform-agnostic frontend-facing DTOs (`LibraryBook`, `LibraryAuthor`,
+# cover/file URLs) and the `libcalibre` -> DTO conversion. Depends on neither
+# Tauri nor any HTTP framework: the URL scheme is injected by the caller, so
+# both the Tauri app (`asset://`) and a future `citadel-server` (`https://`)
+# can build the same shapes. See AGENTS.md.
+
+[dependencies]
+libcalibre = { path = "../libcalibre" }
+serde = { version = "1.0", features = ["derive"] }
+# Pinned to src-tauri's specta so the `Type` derive emits identical bindings.
+specta = { version = "=2.0.0-rc.22", features = ["derive"] }
+urlencoding = "2.1.3"
+
+[dev-dependencies]
+chrono = { version = "0.4.31", features = ["serde"] }
