+++ src-tauri/Cargo.toml
@@ -19,6 +19,7 @@ chrono = { version = "0.4.31", features = ["serde"] }
 diesel = { version = "2.1.0", features = ["sqlite", "chrono", "returning_clauses_for_sqlite_3_35"] }
 epub = "2.1.1"
 mobi = "0.8.0"
+citadel-core = { path = "../crates/citadel-core" }
 libcalibre = { path = "../crates/libcalibre" }
 log = "0.4"
 quick-xml = "0.38"
